Refactor MORT Code to be real code, and rudimentary MORT import support
Much more readable decoding code, it's not quite as complex overall as originally thought, and not variable bitrate decoding (that is just the playback engine).  That said, the algorithm is clearly very sophisticated, and if someone wants to make a real importer would be great, as this is a poor brute force algorithm that does not maintain high quality at all.  Still, it's something.